refactor: inline initializeSolidFlowStore into createSolidFlow
