+++ src/data/createSolidFlow.tsx
@@ -1,38 +1,428 @@
 import {
+  BezierEdgeInternal,
+  SmoothStepEdgeInternal,
+  StepEdgeInternal,
+  StraightEdgeInternal,
+} from "@/components/graph/edge";
+import { DefaultNode, GroupNode, InputNode, OutputNode } from "@/components/graph/node";
+import { ReactiveMap } from "@solid-primitives/map";
+import { createMediaQuery } from "@solid-primitives/media";
+import {
   adoptUserNodes,
   calculateNodePosition,
   type Connection,
+  type ConnectionLookup,
+  type ConnectionState,
+  createMarkerIds,
+  devWarn,
+  type EdgeLookup,
   errorMessages,
+  fitViewport,
+  getInternalNodesBounds,
+  getViewportForBounds,
+  type Handle,
+  infiniteExtent,
   initialConnection,
   type InternalNodeBase,
   type InternalNodeUpdate,
+  type MarkerProps,
+  mergeAriaLabelConfig,
   type NodeDimensionChange,
   type NodeDragItem,
+  type NodeLookup,
   type NodePositionChange,
+  type OnError,
   panBy as panBySystem,
+  type PanZoomInstance,
+  type ParentLookup,
+  pointToRendererPoint,
+  type SelectionRect,
   type SetCenterOptions,
   snapPosition,
   addEdge as systemAddEdge,
   updateNodeInternals as systemUpdateNodeInternals,
+  type Transform,
   updateAbsolutePositions,
+  updateConnectionLookup,
+  type Viewport,
   type ViewportHelperFunctionOptions,
   type XYPosition,
 } from "@xyflow/system";
-import { batch, createEffect, createMemo, on } from "solid-js";
+import { batch, createEffect, createMemo, createSignal, mergeProps, on } from "solid-js";
 import { createStore, produce, type SetStoreFunction } from "solid-js/store";
 
 import type { SolidFlowProps } from "@/components/SolidFlow/types";
 import type { FitViewOptions } from "@/shared/types";
-import type { Edge, InternalNode, Node, NodeGraph } from "@/types";
-import { deepTrack } from "@/utils";
+import type {
+  BuiltInEdgeTypes,
+  BuiltInNodeTypes,
+  Edge,
+  EdgeTypes,
+  InternalNode,
+  Node,
+  NodeGraph,
+  NodeTypes,
+} from "@/types";
+import { createWritable, deepTrack } from "@/utils";
 
-import { initializeSolidFlowStore } from "./initializeSolidFlowStore";
+import { getDefaultFlowStateProps } from "./defaults";
+
+type RefinedMarkerProps = Omit<MarkerProps, "markerUnits"> & {
+  readonly markerUnits?: "strokeWidth" | "userSpaceOnUse" | undefined;
+};
+
+export const InitialNodeTypesMap = {
+  input: InputNode,
+  output: OutputNode,
+  default: DefaultNode,
+  group: GroupNode,
+} satisfies BuiltInNodeTypes;
+
+export const InitialEdgeTypesMap = {
+  straight: StraightEdgeInternal,
+  smoothstep: SmoothStepEdgeInternal,
+  default: BezierEdgeInternal,
+  step: StepEdgeInternal,
+} satisfies BuiltInEdgeTypes;
+
+const getInitialViewport = (
+  // This is just used to make sure adoptUserNodes is called before we calculate the viewport
+  _nodesInitialized: boolean,
+  fitView: boolean | undefined,
+  initialViewport: Viewport | undefined,
+  width: number,
+  height: number,
+  nodeLookup: NodeLookup,
+) => {
+  if (fitView && !initialViewport && width && height) {
+    const bounds = getInternalNodesBounds(nodeLookup, {
+      filter: (node) =>
+        !!((node.width || node.initialWidth) && (node.height || node.initialHeight)),
+    });
+    return getViewportForBounds(bounds, width, height, 0.5, 2, 0.1);
+  } else {
+    return initialViewport ?? { x: 0, y: 0, zoom: 1 };
+  }
+};
 
 export const createSolidFlow = <NodeType extends Node = Node, EdgeType extends Edge = Edge>(
   props: Partial<SolidFlowProps<NodeType, EdgeType>>,
 ) => {
-  const solidFlowStore = initializeSolidFlowStore(props);
-  const { store, nodeLookup, parentLookup, edgeLookup, connectionLookup, actions } = solidFlowStore;
+  const _props = mergeProps(getDefaultFlowStateProps<NodeType, EdgeType>(), props);
+
+  const nodeLookup: NodeLookup<InternalNode<NodeType>> = new ReactiveMap();
+  const parentLookup: ParentLookup<InternalNode<NodeType>> = new ReactiveMap();
+  const edgeLookup: EdgeLookup<EdgeType> = new ReactiveMap();
+  const connectionLookup: ConnectionLookup = new ReactiveMap();
+
+  batch(() => {
+    updateConnectionLookup(connectionLookup, edgeLookup, _props.edges);
+  });
+
+  const startNodesInitialized = batch(() => {
+    return adoptUserNodes(_props.nodes, nodeLookup, parentLookup, {
+      nodeExtent: _props.nodeExtent,
+      nodeOrigin: _props.nodeOrigin,
+      elevateNodesOnSelect: _props.elevateNodesOnSelect,
+      checkEquality: true,
+    });
+  });
+
+  const _viewport = getInitialViewport(
+    startNodesInitialized,
+    _props.fitView,
+    _props.initialViewport,
+    _props.width ?? 0,
+    _props.height ?? 0,
+    nodeLookup,
+  );
+
+  const mediaPrefersDark = createMediaQuery(
+    "(prefers-color-scheme: dark)",
+    _props.colorModeSSR === "dark",
+  );
+
+  const [config, setConfig] = createSignal(_props);
+
+  const [ariaLabelConfig, setAriaLabelConfig] = createWritable(() =>
+    mergeAriaLabelConfig(config().ariaLabelConfig),
+  );
+  const [ariaLiveMessage, setAriaLiveMessage] = createWritable(() => config().ariaLiveMessage);
+  const [clickConnectStartHandle, setClickConnectStartHandle] = createSignal<
+    Pick<Handle, "id" | "nodeId" | "type"> | undefined
+  >(undefined);
+  const [connection, setConnection] =
+    createSignal<ConnectionState<InternalNode<NodeType>>>(initialConnection);
+  const [domNode, setDomNode] = createSignal<HTMLElement | null>(null);
+  const [dragging, setDragging] = createSignal(false);
+  const [fitViewResolver, setFitViewResolver] = createSignal<
+    PromiseWithResolvers<boolean> | undefined
+  >();
+  const [elementsSelectable, setElementsSelectable] = createWritable(
+    () => config().elementsSelectable,
+  );
+  const [fitViewQueued, setFitViewQueued] = createWritable(() => config().fitViewQueued);
+  const [fitViewOptions, setFitViewOptions] = createWritable(() => config().fitViewOptions);
+  const [height, setHeight] = createWritable(() => config().height);
+  const [minZoom, _setMinZoom] = createWritable<number>(() => config().minZoom);
+  const [maxZoom, _setMaxZoom] = createWritable<number>(() => config().maxZoom);
+  const [nodesConnectable, setNodesConnectable] = createWritable(() => config().nodesDraggable);
+  const [nodesDraggable, setNodesDraggable] = createWritable(() => config().nodesDraggable);
+  const [panZoom, setPanZoom] = createSignal<PanZoomInstance | null>(null);
+  const [selectionRect, setSelectionRect] = createSignal<SelectionRect | undefined>();
+  const [selectionRectMode, setSelectionRectMode] = createSignal<string | undefined>();
+  const [snapGrid, setSnapGrid] = createWritable(() => config().snapGrid);
+  const [translateExtent, _setTranslateExtent] = createWritable(
+    () => config().translateExtent ?? infiniteExtent,
+  );
+  const [viewport, setViewport] = createSignal<Viewport>(_viewport);
+  const [width, setWidth] = createWritable(() => config().width);
+
+  // Key flags
+  const [selectionKeyPressed, setSelectionKeyPressed] = createSignal(false);
+  const [multiselectionKeyPressed, setMultiselectionKeyPressed] = createSignal(false);
+  const [deleteKeyPressed, setDeleteKeyPressed] = createSignal(false);
+  const [panActivationKeyPressed, setPanActivationKeyPressed] = createSignal(false);
+  const [zoomActivationKeyPressed, setZoomActivationKeyPressed] = createSignal(false);
+
+  createEffect(() => {
+    const _panZoom = panZoom();
+    if (!_panZoom) return;
+    createEffect(() => {
+      _panZoom.setScaleExtent([store.minZoom, store.maxZoom]);
+    });
+    createEffect(() => {
+      _panZoom.setTranslateExtent(store.translateExtent);
+    });
+  });
+
+  const store = mergeProps({ width: 0, height: 0 }, config, {
+    get _colorMode() {
+      return config().colorMode;
+    },
+    get _colorModeSSR() {
+      return config().colorModeSSR;
+    },
+    get _connection() {
+      return connection();
+    },
+    get _nodeTypes() {
+      return config().nodeTypes;
+    },
+    get _edgeTypes() {
+      return config().edgeTypes;
+    },
+    get ariaLabelConfig() {
+      return ariaLabelConfig();
+    },
+    get ariaLiveMessage() {
+      return ariaLiveMessage();
+    },
+    get clickConnectStartHandle() {
+      return clickConnectStartHandle();
+    },
+    get colorMode() {
+      return this._colorMode === "system"
+        ? mediaPrefersDark()
+          ? "dark"
+          : "light"
+        : this._colorMode;
+    },
+    get connection() {
+      const state = connection();
+      return {
+        ...state,
+        to: state.inProgress ? pointToRendererPoint(state.to, this.transform) : state.to,
+      } as ConnectionState<InternalNode<NodeType>>;
+    },
+    get domNode() {
+      return domNode();
+    },
+    get dragging() {
+      return dragging();
+    },
+    get edgeTypes() {
+      return { ...InitialEdgeTypesMap, ...this._edgeTypes } as EdgeTypes;
+    },
+    get elementsSelectable() {
+      return elementsSelectable();
+    },
+    get fitViewResolver() {
+      return fitViewResolver();
+    },
+    get fitViewQueued() {
+      return fitViewQueued();
+    },
+    get fitViewOptions() {
+      return fitViewOptions();
+    },
+    get height() {
+      return height();
+    },
+    get lib() {
+      /*
+       * Made this a derived store get value to prevent overwriting the value. This value is crucial
+       * for the underlying @xyflow/system library to identify elements as this is the prefix we use
+       * for all the CSS class names across the library.
+       */
+      return "solid" as const;
+    },
+    get onError() {
+      return config().onFlowError ?? (devWarn as OnError);
+    },
+    get markers() {
+      return createMarkerIds(config().edges, {
+        defaultColor: config().defaultMarkerColor,
+        id: config().id,
+      }) as RefinedMarkerProps[];
+    },
+    get maxZoom() {
+      return maxZoom();
+    },
+    get minZoom() {
+      return minZoom();
+    },
+    get nodesConnectable() {
+      return nodesConnectable();
+    },
+    get nodesDraggable() {
+      return nodesDraggable();
+    },
+    get nodeTypes() {
+      return { ...InitialNodeTypesMap, ...this._nodeTypes } as NodeTypes;
+    },
+    get panZoom() {
+      return panZoom();
+    },
+    get selectedNodes() {
+      return config().nodes.filter((node) => node.selected);
+    },
+    get selectedEdges() {
+      return config().edges.filter((edge) => edge.selected);
+    },
+    get selectionRect() {
+      return selectionRect();
+    },
+    get selectionRectMode() {
+      return selectionRectMode();
+    },
+    get snapGrid() {
+      return snapGrid();
+    },
+    get viewport() {
+      return viewport();
+    },
+    get viewportInitialized() {
+      return panZoom() !== null;
+    },
+    get transform() {
+      return [viewport().x, viewport().y, viewport().zoom] as Transform;
+    },
+    get translateExtent() {
+      return translateExtent();
+    },
+    get width() {
+      return width();
+    },
+
+    // key press flags
+    get selectionKeyPressed() {
+      return selectionKeyPressed();
+    },
+    get multiselectionKeyPressed() {
+      return multiselectionKeyPressed();
+    },
+    get deleteKeyPressed() {
+      return deleteKeyPressed();
+    },
+    get panActivationKeyPressed() {
+      return panActivationKeyPressed();
+    },
+    get zoomActivationKeyPressed() {
+      return zoomActivationKeyPressed();
+    },
+  });
+
+  const resolveFitView = async () => {
+    if (!store.panZoom) return;
+
+    await fitViewport(
+      {
+        nodes: nodeLookup,
+        width: store.width,
+        height: store.height,
+        panZoom: store.panZoom,
+        minZoom: store.minZoom,
+        maxZoom: store.maxZoom,
+      },
+      fitViewOptions(),
+    );
+
+    store.fitViewResolver?.resolve(true);
+
+    /**
+     * wait for the fitViewport to resolve before deleting the resolver,
+     * we want to reuse the old resolver if the user calls fitView again in the mean time
+     */
+    setFitViewQueued(false);
+    setFitViewResolver(undefined);
+    setFitViewOptions(undefined);
+  };
+
+  const updateSystemNodes = () => {
+    return batch(() => {
+      return adoptUserNodes(store.nodes, nodeLookup, parentLookup, {
+        nodeOrigin: store.nodeOrigin,
+        nodeExtent: store.nodeExtent,
+        elevateNodesOnSelect: store.elevateNodesOnSelect,
+        checkEquality: false, // recompute internal nodes
+      });
+    });
+  };
+
+  const nodesInitialized = () => {
+    const result = updateSystemNodes();
+
+    if (store.fitViewQueued) {
+      if (store.fitViewOptions?.duration) {
+        void resolveFitView();
+      } else {
+        queueMicrotask(() => {
+          void resolveFitView();
+        });
+      }
+    }
+
+    return result;
+  };
+
+  const resetStoreValues = () => {
+    // NOTE: should we reset to the config()-values instead?
+    setDragging(false);
+    setSelectionRect(undefined);
+    setSelectionRectMode(undefined);
+    setSelectionKeyPressed(false);
+    setMultiselectionKeyPressed(false);
+    setDeleteKeyPressed(false);
+    setPanActivationKeyPressed(false);
+    setZoomActivationKeyPressed(false);
+    setConnection({ ...initialConnection });
+    setClickConnectStartHandle(undefined);
+    setViewport(config().initialViewport ?? { x: 0, y: 0, zoom: 1 });
+    setAriaLiveMessage("");
+    setSnapGrid(undefined);
+  };
+
+  const _setNodes = createMemo(() => {
+    const [_, set] = createStore(store.nodes);
+    return set;
+  });
+  const _setEdges = createMemo(() => {
+    const [_, set] = createStore(store.edges);
+    return set;
+  });
+  const setNodes: SetStoreFunction<NodeType[]> = (...args) => _setNodes()(...args);
+  const setEdges: SetStoreFunction<Edge[]> = (...args) => _setEdges()(...args);
 
   createEffect(
     on(
@@ -48,27 +438,14 @@ export const createSolidFlow = <NodeType extends Node = Node, EdgeType extends E
     ),
   );
 
-  const _setNodes = createMemo(() => {
-    const [_, set] = createStore(store.nodes);
-    return set;
-  });
-
-  const _setEdges = createMemo(() => {
-    const [_, set] = createStore(store.edges);
-    return set;
-  });
-
-  const setNodes: SetStoreFunction<NodeType[]> = (...args) => _setNodes()(...args);
-  const setEdges: SetStoreFunction<Edge[]> = (...args) => _setEdges()(...args);
-
   const fitView = async (options?: FitViewOptions) => {
     // We either create a new Promise or reuse the existing one
     // Even if fitView is called multiple times in a row, we only end up with a single Promise
     const fitViewResolver = store.fitViewResolver ?? Promise.withResolvers<boolean>();
 
-    actions.setFitViewQueued(true);
-    actions.setFitViewOptions(options);
-    actions.setFitViewResolver(fitViewResolver);
+    setFitViewQueued(true);
+    setFitViewOptions(options);
+    setFitViewResolver(fitViewResolver);
 
     return fitViewResolver.promise;
   };
@@ -109,7 +486,7 @@ export const createSolidFlow = <NodeType extends Node = Node, EdgeType extends E
     });
 
     if (store.fitViewQueued) {
-      void actions.resolveFitView();
+      void resolveFitView();
     }
 
     const nodeToChange = changes.reduce<Map<string, NodeDimensionChange | NodePositionChange>>(
@@ -266,8 +643,8 @@ export const createSolidFlow = <NodeType extends Node = Node, EdgeType extends E
     }
 
     batch(() => {
-      actions.setSelectionRect(undefined);
-      actions.setSelectionRectMode(undefined);
+      setSelectionRect(undefined);
+      setSelectionRectMode(undefined);
     });
 
     if (!node.selected) {
@@ -293,8 +670,8 @@ export const createSolidFlow = <NodeType extends Node = Node, EdgeType extends E
     if (!selectable) return;
 
     batch(() => {
-      actions.setSelectionRect(undefined);
-      actions.setSelectionRectMode(undefined);
+      setSelectionRect(undefined);
+      setSelectionRectMode(undefined);
     });
 
     if (!edge.selected) {
@@ -366,11 +743,11 @@ export const createSolidFlow = <NodeType extends Node = Node, EdgeType extends E
   };
 
   const cancelConnection = () => {
-    actions.setConnection({ ...initialConnection });
+    setConnection({ ...initialConnection });
   };
 
   const reset = () => {
-    actions.resetStoreValues();
+    resetStoreValues();
     unselectNodesAndEdges();
   };
 
@@ -382,10 +759,36 @@ export const createSolidFlow = <NodeType extends Node = Node, EdgeType extends E
     parentLookup,
     connectionLookup,
     actions: {
-      ...actions,
-
-      setNodes,
+      nodesInitialized,
+      resetStoreValues,
+      resolveFitView,
+      setAriaLabelConfig,
+      setAriaLiveMessage,
+      setClickConnectStartHandle,
+      setConfig,
+      setConnection,
+      setDeleteKeyPressed,
+      setDomNode,
+      setDragging,
       setEdges,
+      setElementsSelectable,
+      setFitViewOptions,
+      setFitViewQueued,
+      setFitViewResolver,
+      setHeight,
+      setMultiselectionKeyPressed,
+      setNodes,
+      setNodesConnectable,
+      setNodesDraggable,
+      setPanActivationKeyPressed,
+      setPanZoom,
+      setSelectionKeyPressed,
+      setSelectionRect,
+      setSelectionRectMode,
+      setViewport,
+      setWidth,
+      setZoomActivationKeyPressed,
+      updateSystemNodes,
 
       // Port Svelte Flow Actions to Solid Flow
       addEdge,
